Turned up gcc warnings, fixed warnings in code, enabled -Werror. Note: #include "qpid_test_plugin.h" instead of <cppunit/TestPlugin.h> Works around warning from a cppunit macro.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk@462834 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/common/io/inc/APRConnector.h
@@ -68,7 +68,7 @@ namespace io {
 
         void checkIdle(apr_status_t status);
 	void writeBlock(qpid::framing::AMQDataBlock* data);
-	void writeToSocket(char* data, int available);
+	void writeToSocket(char* data, size_t available);
         void setSocketTimeout();
 
 	void run();
